+++ react-crud/src/App.js
@@ -1,25 +1,22 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
-import Navbar from "./components/Navbar/Navbar";
 import Home from "./components/Home/home";
-import Update from "./components/Update";
-import Create from "./components/Create";
-import Search from "./components/Search";
+import Update from "./components/Update/Update";
+import Create from "./components/Create/Create";
+import Search from "./components/Search/Search";
 import Login from "./components/Login/Login";
+import Navbar from "./components/Navbar/Navbar";
 
 function App() {
   return (
     <Router>
-      <Navbar />
-      <div className="main">
-        <Routes>
-          <Route exact path="/" Component={Login} />
-          <Route exact path="/home" Component={Home} />
-          <Route exact path="/update" Component={Update} />
-          <Route exact path="/create" Component={Create} />
-          <Route exact path="/read" Component={Search} />
-        </Routes>
-      </div>
+      <Routes>
+        <Route exact path="/" Component={Login} />
+        <Route exact path="/home" Component={Home} />
+        <Route exact path="/update" Component={Update} />
+        <Route exact path="/create" Component={Create} />
+        <Route exact path="/read" Component={Search} />
+      </Routes>
     </Router>
   );
 }
